refactor(huds): cambia nombre al modulo
